+++ pages/api/deleteUnauthorized.ts
@@ -1,6 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 
-import DeleteUnauthorized from '../../src/graphql/resolvers/User/Mutation/DeleteUnauthorized';
+import DeleteUnauthorized from '../../src/graphql/resolvers/user/mutations/DeleteUnauthorized';
 
 const Handler = async (req: NextApiRequest, res: NextApiResponse) => {
   res.status(200).json({ success: await DeleteUnauthorized() });
