+++ src/graphql/resolvers/collection/DeleteCollection.ts
