+++ src/graphql/resolvers/collection/RemovePostFromCollection.ts
