+++ src/graphql/resolvers/index.ts
@@ -1,27 +1,27 @@
-export { default as AddPostToCollection } from './Collection/AddPostToCollection';
-export { default as CollectionQuery } from './Collection/CollectionQuery';
-export { default as CreateCollection } from './Collection/CreateCollection';
-export { default as EditCollection } from './Collection/EditCollection';
-export { default as RemovePostFromCollection } from './Collection/RemovePostFromCollection';
-export { default as CreateComment } from './Comment/CreateComment';
-export { default as Delete } from './Delete/Delete';
-export { default as PlaceQuery } from './Place/PlaceQuery';
-export { default as FilterPlacesQuery } from './Place/PlacesQuery';
-export { default as CreatePost } from './Post/CreatePost';
-export { default as PersonalizedPostsQuery } from './Post/PersonalizedPostsQuery';
-export { default as PostQuery } from './Post/PostQuery';
-export { default as PostsQuery } from './Post/PostsQuery';
-export { default as Follow } from './Relations/Follow';
-export { default as Like } from './Relations/Like';
-export { default as Report } from './Relations/Report';
-export { default as Unfollow } from './Relations/Unfollow';
-export { default as Unlike } from './Relations/Unlike';
-export { default as EditUser } from './User/Mutation/EditUser';
-export { default as GoogleAuth } from './User/Mutation/GoogleAuth';
-export { default as Login } from './User/Query/Login';
-export { default as CreateUser } from './User/Mutation/CreateUser';
-export { default as DeleteUser } from './User/Mutation/DeleteUser';
-export { default as MeQuery } from './User/Query/MeQuery';
-export { default as SuggestedUsersQuery } from './User/Query/SuggestedUsersQuery';
-export { default as UserQuery } from './User/Query/UserQuery';
-export { default as VerifyToken } from './User/Mutation/VerifyToken';
+export { default as AddPostToCollection } from './collection/AddPostToCollection';
+export { default as CollectionQuery } from './collection/CollectionQuery';
+export { default as CreateCollection } from './collection/CreateCollection';
+export { default as EditCollection } from './collection/EditCollection';
+export { default as RemovePostFromCollection } from './collection/RemovePostFromCollection';
+export { default as CreateComment } from './comment/CreateComment';
+export { default as Delete } from './delete/Delete';
+export { default as PlaceQuery } from './place/PlaceQuery';
+export { default as FilterPlacesQuery } from './place/PlacesQuery';
+export { default as CreatePost } from './post/CreatePost';
+export { default as PersonalizedPostsQuery } from './post/PersonalizedPostsQuery';
+export { default as PostQuery } from './post/PostQuery';
+export { default as PostsQuery } from './post/PostsQuery';
+export { default as Follow } from './relations/Follow';
+export { default as Like } from './relations/Like';
+export { default as Report } from './relations/Report';
+export { default as Unfollow } from './relations/Unfollow';
+export { default as Unlike } from './relations/Unlike';
+export { default as EditUser } from './user/mutations/EditUser';
+export { default as GoogleAuth } from './user/mutations/GoogleAuth';
+export { default as Login } from './user/queries/Login';
+export { default as CreateUser } from './user/mutations/CreateUser';
+export { default as DeleteUser } from './user/mutations/DeleteUser';
+export { default as MeQuery } from './user/queries/MeQuery';
+export { default as SuggestedUsersQuery } from './user/queries/SuggestedUsersQuery';
+export { default as UserQuery } from './user/queries/UserQuery';
+export { default as VerifyToken } from './user/mutations/VerifyToken';
