+++ src/graphql/resolvers/search/FullTextSearch.ts
