+++ src/graphql/resolvers/user/mutations/DeleteUnauthorized.ts
