+++ src/graphql/resolvers/user/queries/SuggestedUsersQuery.ts
